pyproject.toml
[project]
name = "graphql_db"
version = "0.1.0"
description = "GraphQL Database Mapper - Generate GraphQL APIs from database models (SQLAlchemy, SQLModel, etc.)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Robert Parker", email = "rob@parob.com"}
]
keywords = ["GraphQL", "GraphQL API", "Database", "ORM", "SQLAlchemy", "SQLModel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "graphql-core>=3.0.0",
    "sqlalchemy>=2.0.0",
    "sqlalchemy-utils>=0.40.0",
    "graphql-api>=1.0.2",
    "context-helper>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "coverage[toml]>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "sqlmodel>=0.0.14",
]

[project.urls]
Homepage = "https://gitlab.com/parob/graphql-db"
Repository = "https://gitlab.com/parob/graphql-db"
"Bug Reports" = "https://gitlab.com/parob/graphql-db/-/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Version is now static in [project] section

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["S101", "N806", "N803"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=graphql_db --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["graphql_db"]
omit = ["tests/*", "**/_version.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.1",
    "sqlmodel>=0.0.25",
]
